+++ django_comments_tree/conf/defaults.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals
 from django.conf import settings
 import markdown
-from django_comments_tree.render import render_draftjs
+from django_comments_tree.render import render_draftjs, render_plain
 
 COMMENT_MAX_LENGTH = 3000
 
@@ -67,6 +67,7 @@ COMMENTS_TREE_FIREBASE_KEY = None
 
 # Default types we can use for comments
 MARKUP_FIELD_TYPES = (
+    ('plain', render_plain),
     ('markdown', markdown.markdown),
-    ('Draft.js', render_draftjs),
+    ('draftjs', render_draftjs),
 )